# swagger.yaml
openapi: 3.0.3
info:
  title: SET Service API
  version: 1.1.0
  description: |
    SET Service REST API for Worker, Company, Super Admin, and restricted Admin roles.

    Core worker flow:
    Worker registers -> verifies OTP -> becomes pending_approval -> super admin or permitted admin approves or rejects -> approved worker can log in and receive tokens.

    Contract-wide observability and authentication rules:
    - Every HTTP response includes `X-Request-ID` and `X-Correlation-ID`.
    - Clients may send either header using 8-128 characters from `[A-Za-z0-9._:-]`; invalid values are replaced server-side.
    - Error bodies include the effective `request_id`; the correlation identifier remains in the response header.
    - `BearerAuth` accepts access tokens only. Refresh tokens are valid only in the documented refresh/logout request bodies.
    - OTP request success means the request was accepted for delivery; it does not prove arrival at the handset or mailbox.
    - Interactive Swagger UI is disabled in production and should be exposed only in controlled non-production environments.
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://api.setservice.az/v1
    description: Production
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT access token
      description: Short-lived SET Service JWT access token. Refresh tokens are rejected by authenticated API middleware.
...
  schemas:
    RequestId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
      example: 7fa22d18-6e54-4c33-8154-9f68f8e2f84b
    Error:
      type: object
      required: [error, code, request_id, timestamp]
      properties:
        error: { type: string, example: Worker is not approved for login. }
        code: { type: string, example: WORKER_NOT_APPROVED }
        details: { type: object, nullable: true, example: { status: pending_approval } }
        request_id: { $ref: '#/components/schemas/RequestId' }
        timestamp: { type: string, format: date-time }
...
    AdminPermission:
      type: string
      description: Manage permissions are normalized with required view permissions. `manage_workers` implies `view_workers`; `manage_companies` implies `view_companies`; `manage_assignments` implies `view_assignments`, `view_orders`, and `view_workers`; `manage_kiosks` implies `view_orders` and `view_assignments`.
      enum:
        - view_dashboard
        - view_workers
        - manage_workers
        - view_companies
        - manage_companies
        - view_orders
        - view_assignments
        - manage_assignments
        - view_attendance
        - view_reports
        - manage_kiosks
        - view_notifications
        - manage_admins
    ManagedAdmin:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        name: { type: string, example: Operations Admin }
        email: { type: string, format: email, example: ops@setservice.az }
        role: { type: string, enum: [admin] }
        is_active: { type: boolean, example: true }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/AdminPermission' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateAdminRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, example: Operations Admin }
        email: { type: string, format: email, example: ops@setservice.az }
        password: { type: string, minLength: 8, example: <explicit-admin-password> }
        is_active: { type: boolean, default: true }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/AdminPermission' }
    UpdateAdminRequest:
      type: object
      properties:
        name: { type: string, example: Operations Admin }
        email: { type: string, format: email, example: ops@setservice.az }
        password: { type: string, minLength: 8, example: <explicit-admin-password> }
        is_active: { type: boolean }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/AdminPermission' }
...
paths:
...
  /admin/admins:
    get:
      tags: [Admin - Admins]
      security: [{ BearerAuth: [] }]
      summary: List restricted admins
      description: Super admin bypasses permissions. Restricted admins need `manage_admins`.
      responses:
        '200':
          description: Admin list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ManagedAdmin' }
        '403': { description: Permission denied, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      tags: [Admin - Admins]
      security: [{ BearerAuth: [] }]
      summary: Create restricted admin
      description: No public admin signup exists. Super admin or an admin with `manage_admins` creates admin accounts.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAdminRequest' }
      responses:
        '201': { description: Admin created, content: { application/json: { schema: { $ref: '#/components/schemas/ManagedAdmin' } } } }
        '403': { description: Permission denied, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Email already exists, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /admin/admins/{id}:
    patch:
      tags: [Admin - Admins]
      security: [{ BearerAuth: [] }]
      summary: Update restricted admin
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAdminRequest' }
      responses:
        '200': { description: Admin updated, content: { application/json: { schema: { $ref: '#/components/schemas/ManagedAdmin' } } } }
        '403': { description: Permission denied, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Admin not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [Admin - Admins]
      security: [{ BearerAuth: [] }]
      summary: Deactivate restricted admin
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Admin deactivated, content: { application/json: { schema: { $ref: '#/components/schemas/ManagedAdmin' } } } }
        '403': { description: Permission denied, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /taxonomy/admin:
    get:
      tags: [Taxonomy]
      security: [{ BearerAuth: [] }]
      summary: List active and inactive taxonomy
      description: Requires super_admin or admin with manage_admins.
      responses:
        '200': { description: Full taxonomy tree }
        '403': { description: Missing manage_admins permission }
...